refactor: 클래스 등록 form을 useClassRegisterForm 커스텀 훅으로 분리 (#142)
